+++ packages/reactive_ble_mobile/android/src/main/kotlin/com/signify/hue/flutterreactiveble/ble/ReactiveBleClient.kt
@@ -1127,4 +1127,8 @@ open class ReactiveBleClient(private val context: Context) : BleClient {
                 .setShouldLogAttributeValues(true)
                 .build()
         )
+
+    fun isDeviceConnected(deviceId: String): Observable<EstablishConnectionResult> {
+      return getConnection(deviceId)
+    }
 }